Split out room
Move the existing `Room` class to its own file.

I've made one structural change: the only user of the method `fetch_events` is the `Room` class, and it refers to a specific room, so I've turned it into an instance method of `Room`.
